requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-sqlalchemy==3.1.1
marshmallow==3.21.0
python-dotenv==1.0.1
email-validator==2.1.1
